Remove the TreeStorage interfaces (#2586)
This change removes the unnecessary TreeStorage interfaces now that
there is only log storage after maps removal. All the methods are still
embedded in LogStorage interfaces, implementations shouldn't notice
this transition (may only need minor type renames similarly to how it is
done in this PR).
